+++ CHANGELOG.md
@@ -151,6 +151,7 @@
 * [X] breeze-core 提供字符串的常量集。
 * [X] 移除breeze-fxgl（考虑使用LibGDX）。
 * [X] breeze-dsl 完善`CommandLineTextDsl`，以支持更多的富文本。
+* [X] breeze-dsl 实现`MermaidPieChartDsl`。
 * [ ] breeze-dsl 实现`YamlDsl`。（不推荐使用，但是作为规范）
 
 ***